move auth installables to separate namespace
