+++ src/Tempest/Auth/src/Install/PermissionMigration.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tempest\Auth;
+namespace Tempest\Auth\Install;
 
 use Tempest\Core\DoNotDiscover;
 use Tempest\Database\Migration;
@@ -10,19 +10,18 @@ use Tempest\Database\QueryStatements\CreateTableStatement;
 use Tempest\Database\QueryStatements\DropTableStatement;
 
 #[DoNotDiscover]
-final readonly class UserPermissionMigration implements Migration
+final readonly class PermissionMigration implements Migration
 {
     public function getName(): string
     {
-        return '0000-00-02_create_user_permissions_table';
+        return '0000-00-01_create_permissions_table';
     }
 
     public function up(): CreateTableStatement
     {
-        return (new CreateTableStatement('user_permissions'))
+        return (new CreateTableStatement('permissions'))
             ->primary()
-            ->belongsTo('user_permissions.user_id', 'users.id')
-            ->belongsTo('user_permissions.permission_id', 'permissions.id');
+            ->varchar('name');
     }
 
     public function down(): DropTableStatement
