+++ src/Tempest/Auth/src/Install/UserMigration.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tempest\Auth;
+namespace Tempest\Auth\Install;
 
 use Tempest\Core\DoNotDiscover;
 use Tempest\Database\Migration;
@@ -10,22 +10,25 @@ use Tempest\Database\QueryStatements\CreateTableStatement;
 use Tempest\Database\QueryStatements\DropTableStatement;
 
 #[DoNotDiscover]
-final readonly class PermissionMigration implements Migration
+final readonly class UserMigration implements Migration
 {
     public function getName(): string
     {
-        return '0000-00-01_create_permissions_table';
+        return '0000-00-00_create_users_table';
     }
 
     public function up(): CreateTableStatement
     {
-        return (new CreateTableStatement('permissions'))
+        return (new CreateTableStatement('users'))
             ->primary()
-            ->varchar('name');
+            ->varchar('name')
+            ->varchar('email')
+            ->datetime('emailValidatedAt', nullable: true)
+            ->text('password');
     }
 
     public function down(): DropTableStatement
     {
-        return DropTableStatement::forModel(Permission::class);
+        return DropTableStatement::forModel(User::class);
     }
 }
